sittings: add a drop-down list to choose committees
* $state.params.sitting is the ID of displaying sitting.
* Context is the type of displaying committee.

any modification of $state.params.sitting will trigger
checking to see whether the context should be changed.

each time we change the context, it fetch the list of corresponding
committees and stores into $scope.lists then use it as current list.

each time the current list changed, we pick one sitting of the list
to load and display.